NEW: Lagt til AddOne for å holde status riktig på WorkDone. TWEAK: Bruker double på PercentDone. BUGFIX: Rettet problem med slettede rader i konfliktlisten.

+++ WebApp/Interfaces/IDataContext.cs
@@ -13,7 +13,7 @@ namespace TranslateWebApp.Interfaces
         Task<Disagreements> GetDisagreements(int projectId, string langCode);
 
         public int ProjectId { get; }
-        public int PercentDone();
+        public double PercentDone();
         Task DiscardBlock(int blockId);
 
         public string HelperLanguage { get; set; }
